+++ packages/core/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## master
 
+- Added AnyCable signed streams support. ([@palkan][])
+
+  Two new methods have been added to connect to streams directly without any channels: `cable.streamFrom(name)` and `cable.streamFromSigned(signedName)`. See [signed streams docs](https://docs.anycable.io/edge/anycable-go/signed_streams).
+
 ## 0.7.13 (2024-02-26)
 
 - Do not try to send `pong` if cable is no longer connected. ([@palkan][])